Modified line_items controller, lne_item model. Minor view chahges

+++ db/migrate/20191017084856_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
   def change
     create_table :products do |t|
       t.string :name, null: false
-      t.decimal :price, default: 0.0
+      t.float :price, default: 0.0
       t.timestamps
     end
   end
